+++ CHANGELOG.md
@@ -3,6 +3,16 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
+## 2.60.0 - 2023-05-16
+### Added
+- Support for self-service integration in the Fusion Apps as a Service service
+ 
+### Breaking Changes
+- The `AttachExistingInstanceDetails`, `CreateNewInstanceDetails`, `CreateOaxServiceInstanceDetails`, `CreateOicServiceInstanceDetails`, and `CreateServiceInstanceDetails` models were removed in the Fusion Apps as a Service service
+- The `action` property was removed in the `CreateServiceAttachmentDetails` model in the Fusion Apps as a Service service- The `action` property was removed in the `CreateServiceAttachmentDetails` model in the Fusion Apps as a Service service
+- The `action` property was removed in the `ServiceAttachment` model in the Fusion Apps as a Service service
+
+
 ## 2.59.1 - 2023-05-09
 ### Added
 - Support for the Access Governance service